Add a splunk-ready instance
This creates a SplunkReadyInstance.

It has two network interfaces: one is in the Admin subnet (next to the
BIG-IP admin interface).  The other is in the private subnet (next to
the Mesos worker nodes).

Added an output so the user can see the ssh command they need to get
started.

Since there are now several instances running around, added some "Name"
labels so the EC2 instances view is more useful.