Update the CFT and guide to tell you to use the private IP.
If you use the public IP, it only works if you didn't specify a CIDR for
filtering admin access (or you go and add rules).  If you use the
private IP, it will work.

It's easiest to just use the admin private IP because it's routable from
the internal net (mesos nodes) and also BIG-IP's admin interface.  So
remove the SplunkReadyPrivateXface entirely to keep things simple.

Also add an output "SplunkReadyPrivateIP" to the CFT so that users don't
have to go digging through EC2 to find the IP.