Fixes #31: todos still left
Problem: 3 todos still remain in the docs.

    $ grep todo $(find ./ -name "*.rst")
    .//docs/guides/includes/topic_f5-mesos-lwp-install.rst:    .. todo:: add instructions here
    .//docs/guides/lwp_lwp-controller_guide.rst:.. todo:: additional explanation?
    .//usage-marathon-poc.rst:.. todo:: enter email address

Analysis:
For the "add instructions here", copy instructions from the equivalent
f5-marathon-lb section and update.

For the "additional explanation?", what we have is good enough for now.

For the email, confirmed with Sean Duggan that beta@f5.com is the right
email.