Fixes #80 - set up deployment to AWS s3 with secure credentials
Problem: Upon successful builds for release tags, or on commits to any branch that are NOT made by a pull request, we should deploy the built documentation to the correct directory in the AWS s3 bucket.

Analysis: This PR configures travis to handle the s3 deployment. The previous fix for this issue failed to deploy because the access id and key variables were set up incorrectly.