Fix #155: Record that we support BIG-IP 11.6, 13.x in container connectors
The version table shows that the BIG-IP controllers support BIG-IPs
running versions "v12.1.x, v13.0.x" but we also support v11.6, v12.0,
and recommendation from PM is that it would be the most clear to report
this as "v11.6, v12.x, v13.x".

Update the version table for the BIG-IP controllers.  This requires
realigning the other tables but no other content changes made.