Fixes: #166
Problem: The test and deploy scripts use a container image from thejodesterf5. They should use the image from f5devcentral, which is automatically updated when changes are made to the containthedocs project.

Analysis: The test and deploy scripts, and the project README, have been updated as appropriate.

Unit/Robot tests: