Fixes #236: Add docs explaining how f5-kube-proxy and ASP work together
Add a new documentation page explaining the details around f5-kube-proxy
and ASP working together.  Includes example walkthrough that I used to
generate the iptables rules.  This starts with a non-ASP deployment and
then shows the changes that happen when we use f5-kube-proxy and ASP.

Mention the versions I used to create it but indicate that the details
around the rules are an implementation detail - we're providing them as
an example.

Simplify iptables output in example

The iptables output contained many rules but this can make it hard to
focus on what's important.  Reduce this output.  I'm leaving this as a
separate commit so that we can revert and add more back later if we
decide that we removed too much.