add iapp-traffic-group clarification
Problem:
Traffic group configuration for the BIG-IP and partition managed by the k8s-bigip-ctlr could lead to an instance where the examples provided would create an iApp not attached to the external net, even though a simpler example of a virtual server would be accessible from the external net.

Solution:
Added details on alternatives for ensuring the virtual IP settings on the BIG IP and the configuration settings in the iApp were correctly aligned.