Ingress IP specifications
With the addition of the new Default IP for Ingress feature, it's important to document the different
ways in which VIPs can be configured for Ingresses.

Documentation has been added to mention the ways in which the IP  can be set, via annotation,
DNS lookup, or the default ip configuration parameter for the controller.