Update VLAN and VXLAN example names
Update the VLAN and VXLAN names for the examples at lines 191 and 208. The one at 191 was incorrect. It had the example name of the overlay VXLAN Tunnel and it needed to be the underlay VLAN for the VXLAN tunnel. On line 208 I updated the text of the example name to matche the exazmple names used later in the doc.